requirements.txt
flask==3.1.0
flask-cors==5.0.1
tensorflow>=2.20.0
Pillow>=11.0.0
numpy>=1.26.0
